+++ x/oracle/app_test.go
@@ -15,7 +15,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
 
-	bandtest "github.com/bandprotocol/chain/v3/app"
+	band "github.com/bandprotocol/chain/v3/app"
 	bandtesting "github.com/bandprotocol/chain/v3/testing"
 	"github.com/bandprotocol/chain/v3/x/oracle/types"
 )
@@ -23,7 +23,7 @@ import (
 type AppTestSuite struct {
 	suite.Suite
 
-	app *bandtest.BandApp
+	app *band.BandApp
 }
 
 func TestAppTestSuite(t *testing.T) {
@@ -32,11 +32,11 @@ func TestAppTestSuite(t *testing.T) {
 
 func (s *AppTestSuite) SetupTest() {
 	dir := testutil.GetTempDir(s.T())
-	s.app = bandtest.SetupWithCustomHome(false, dir)
+	s.app = bandtesting.SetupWithCustomHome(false, dir)
 	ctx := s.app.BaseApp.NewUncachedContext(false, tmproto.Header{})
 
 	// Activate validators
-	for _, v := range bandtest.Validators {
+	for _, v := range bandtesting.Validators {
 		err := s.app.OracleKeeper.Activate(ctx, v.ValAddress)
 		s.Require().NoError(err)
 	}
@@ -56,19 +56,19 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 		2,
 		"app_test",
 		sdk.NewCoins(sdk.NewInt64Coin("uband", 9000000)),
-		bandtest.TestDefaultPrepareGas,
-		bandtest.TestDefaultExecuteGas,
-		bandtest.Validators[0].Address,
+		bandtesting.TestDefaultPrepareGas,
+		bandtesting.TestDefaultExecuteGas,
+		bandtesting.Validators[0].Address,
 	)
 
-	res1 := s.app.AccountKeeper.GetAccount(ctx, bandtest.Validators[0].Address)
+	res1 := s.app.AccountKeeper.GetAccount(ctx, bandtesting.Validators[0].Address)
 	require.NotNil(res1)
 
 	acc1Num := res1.GetAccountNumber()
 	acc1Seq := res1.GetSequence()
 
 	txConfig := moduletestutil.MakeTestTxConfig()
-	_, res, _, err := bandtest.SignCheckDeliver(
+	_, res, _, err := bandtesting.SignCheckDeliver(
 		s.T(),
 		txConfig,
 		s.app.BaseApp,
@@ -79,7 +79,7 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 		[]uint64{acc1Seq},
 		true,
 		true,
-		bandtest.Validators[0].PrivKey,
+		bandtesting.Validators[0].PrivKey,
 	)
 	require.NotNil(res)
 	require.NoError(err)
@@ -88,9 +88,9 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 		types.OracleScriptID(1),
 		[]byte("calldata"),
 		[]sdk.ValAddress{
-			bandtest.Validators[2].ValAddress,
-			bandtest.Validators[0].ValAddress,
-			bandtest.Validators[1].ValAddress,
+			bandtesting.Validators[2].ValAddress,
+			bandtesting.Validators[0].ValAddress,
+			bandtesting.Validators[1].ValAddress,
 		},
 		2,
 		1,
@@ -102,7 +102,7 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 			types.NewRawRequest(3, 3, []byte("test")),
 		},
 		nil,
-		bandtest.TestDefaultExecuteGas,
+		bandtesting.TestDefaultExecuteGas,
 	)
 
 	request, err := s.app.OracleKeeper.GetRequest(ctx, types.RequestID(1))
@@ -115,9 +115,9 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 			types.NewRawReport(2, 0, []byte("answer2")),
 			types.NewRawReport(3, 0, []byte("answer3")),
 		},
-		bandtest.Validators[0].ValAddress,
+		bandtesting.Validators[0].ValAddress,
 	)
-	_, res, _, err = bandtest.SignCheckDeliver(
+	_, res, _, err = bandtesting.SignCheckDeliver(
 		s.T(),
 		txConfig,
 		s.app.BaseApp,
@@ -128,7 +128,7 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 		[]uint64{acc1Seq + 1},
 		true,
 		true,
-		bandtest.Validators[0].PrivKey,
+		bandtesting.Validators[0].PrivKey,
 	)
 	require.NotNil(res)
 	require.NoError(err)
@@ -144,17 +144,17 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 			types.NewRawReport(2, 0, []byte("answer2")),
 			types.NewRawReport(3, 0, []byte("answer3")),
 		},
-		bandtest.Validators[1].ValAddress,
+		bandtesting.Validators[1].ValAddress,
 	)
 
-	res2 := s.app.AccountKeeper.GetAccount(ctx, bandtest.Validators[1].Address)
+	res2 := s.app.AccountKeeper.GetAccount(ctx, bandtesting.Validators[1].Address)
 	require.NotNil(res2)
 
 	acc2Num := res2.GetAccountNumber()
 	acc2Seq := res2.GetSequence()
 
 	// res, err = handler(ctx, reportMsg2)
-	_, res, endBlockEvent, err := bandtest.SignCheckDeliver(
+	_, res, endBlockEvent, err := bandtesting.SignCheckDeliver(
 		s.T(),
 		txConfig,
 		s.app.BaseApp,
@@ -165,7 +165,7 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 		[]uint64{acc2Seq},
 		true,
 		true,
-		bandtest.Validators[1].PrivKey,
+		bandtesting.Validators[1].PrivKey,
 	)
 
 	require.NotNil(res)
@@ -218,7 +218,7 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 }
 
 // func TestExpiredRequestOracleData(t *testing.T) {
-// 	app, ctx := bandtest.CreateTestApp(t, true)
+// 	app, ctx := bandtesting.CreateTestApp(t, true)
 // 	k := app.OracleKeeper
 
 // 	ctx = ctx.WithBlockHeight(4).WithBlockTime(time.Unix(1581589790, 0))
@@ -230,9 +230,9 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 // 		2,
 // 		"app_test",
 // 		sdk.NewCoins(sdk.NewCoin("uband", math.NewInt(9000000))),
-// 		bandtest.TestDefaultPrepareGas,
-// 		bandtest.TestDefaultExecuteGas,
-// 		bandtest.Validators[0].Address,
+// 		bandtesting.TestDefaultPrepareGas,
+// 		bandtesting.TestDefaultExecuteGas,
+// 		bandtesting.Validators[0].Address,
 // 	)
 // 	res, err := handler(ctx, requestMsg)
 // 	require.NotNil(t, res)
@@ -242,13 +242,13 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 // 		types.OracleScriptID(1),
 // 		[]byte("calldata"),
 // 		[]sdk.ValAddress{
-// 			bandtest.Validators[2].ValAddress,
-// 			bandtest.Validators[0].ValAddress,
-// 			bandtest.Validators[1].ValAddress,
+// 			bandtesting.Validators[2].ValAddress,
+// 			bandtesting.Validators[0].ValAddress,
+// 			bandtesting.Validators[1].ValAddress,
 // 		},
 // 		2,
 // 		4,
-// 		bandtest.ParseTime(1581589790),
+// 		bandtesting.ParseTime(1581589790),
 // 		"app_test",
 // 		[]types.RawRequest{
 // 			types.NewRawRequest(1, 1, []byte("test")),
@@ -256,7 +256,7 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 // 			types.NewRawRequest(3, 3, []byte("test")),
 // 		},
 // 		nil,
-// 		bandtest.TestDefaultExecuteGas,
+// 		bandtesting.TestDefaultExecuteGas,
 // 	)
 // 	app.EndBlocker(ctx.WithBlockHeight(4))
 // 	request, err := k.GetRequest(ctx, types.RequestID(1))
@@ -283,7 +283,7 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 // 		Attributes: []abci.EventAttribute{
 // 			{
 // 				Key:   types.AttributeKeyValidator,
-// 				Value: fmt.Sprint(bandtest.Validators[2].ValAddress.String()),
+// 				Value: fmt.Sprint(bandtesting.Validators[2].ValAddress.String()),
 // 			},
 // 		},
 // 	}, {
@@ -291,7 +291,7 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 // 		Attributes: []abci.EventAttribute{
 // 			{
 // 				Key:   types.AttributeKeyValidator,
-// 				Value: fmt.Sprint(bandtest.Validators[0].ValAddress.String()),
+// 				Value: fmt.Sprint(bandtesting.Validators[0].ValAddress.String()),
 // 			},
 // 		},
 // 	}, {
@@ -299,7 +299,7 @@ func (s *AppTestSuite) TestSuccessRequestOracleData() {
 // 		Attributes: []abci.EventAttribute{
 // 			{
 // 				Key:   types.AttributeKeyValidator,
-// 				Value: fmt.Sprint(bandtest.Validators[1].ValAddress.String()),
+// 				Value: fmt.Sprint(bandtesting.Validators[1].ValAddress.String()),
 // 			},
 // 		},
 // 	}}
